reviewed the issue#10275
